requirements.txt
playwright==1.44.0
playwright-stealth==1.0.6
groq==0.9.0
gspread==6.1.2
google-auth==2.29.0
pdfplumber==0.11.0
python-docx==1.1.2
python-dotenv==1.0.1
tqdm==4.66.4
fastapi==0.111.0
uvicorn==0.30.1
dataclasses-json==0.6.7
